auth_api/python/requirements.txt
Flask
pyjwt
mysql-connector-python